Divide ngraph_var and ngraph_find_replace_prefetchdataset into .h and .cc respectively
