+++ ngraph_bridge/enable_variable_ops/ngraph_var.h
@@ -40,115 +40,39 @@ namespace ngraph_bridge {
 // THIS CLASS IS NOT BEING USED ANYWHERE
 class NGraphVar : public ResourceBase {
  public:
-  explicit NGraphVar(DataType dtype, TensorShape shape, string BackendName)
-      : tf_tensor_(dtype, shape), ng_backend_name_(BackendName) {
-    // TF datatype to nGraph element type
-    ng::element::Type ng_element_type;
-    TFDataTypeToNGraphElementType(dtype, &ng_element_type);
+  explicit NGraphVar(DataType dtype, TensorShape shape, string BackendName);
 
-    // TF TensorShape to nGraphShape
-    ng::Shape ng_shape(shape.dims());
-    for (int j = 0; j < shape.dims(); ++j) {
-      ng_shape[j] = shape.dim_size(j);
-    }
-
-    // Create Backend
-    NGRAPH_VLOG(4) << "NGraphVar::Create Backend ";
-    Status status;
-    ng::runtime::Backend* op_backend;
-    status = BackendManager::CreateBackend(ng_backend_name_);
-    if (!status.ok()) {
-      throw std::runtime_error("Cannot create backend " + ng_backend_name_ +
-                               ". Got Exception " + status.error_message());
-    }
-
-    try {
-      op_backend = BackendManager::GetBackend(ng_backend_name_);
-    } catch (...) {
-      throw std::runtime_error("No backend available :" + ng_backend_name_ +
-                               ". Cannot execute graph");
-    }
-
-    // Create nGTensor
-
-    // Check buffer sharing
-    // -1 implies env var is not set
-    int buffer_sharing_state_env = -1;
-    // 0 implies buffer sharing is disabled
-    // 1 implies buffer sharing is enabled
-    status = GetNgraphVarBufferSharingState(buffer_sharing_state_env);
-    if (!status.ok()) {
-      throw std::runtime_error("Got Exception " + status.error_message());
-    }
-
-    ng_tf_share_buffer_ = (buffer_sharing_state_env == -1)
-                              ? (ng_backend_name_ == "CPU")
-                              : buffer_sharing_state_env;
-
-    if (ng_tf_share_buffer_) {
-      void* tf_src_ptr = (void*)DMAHelper::base(&tf_tensor_);
-      ng_tensor_ =
-          op_backend->create_tensor(ng_element_type, ng_shape, tf_src_ptr);
-    } else {
-      ng_tensor_ = op_backend->create_tensor(ng_element_type, ng_shape);
-    }
-  }
   // Not copyable or movable.
   NGraphVar(const NGraphVar&) = delete;
   NGraphVar& operator=(const NGraphVar&) = delete;
 
-  mutex* mu() { return &mu_; }
-  Tensor* tensor() { return &tf_tensor_; }
-  shared_ptr<ngraph::runtime::Tensor> ng_tensor() { return ng_tensor_; };
+  mutex* mu();
+  Tensor* tensor();
+  shared_ptr<ngraph::runtime::Tensor> ng_tensor();
 
-  string DebugString() const override {
-    return strings::StrCat(DataTypeString(tf_tensor_.dtype()), "/",
-                           tf_tensor_.shape().DebugString());
-  }
+  string DebugString() const override;
 
   // Copies the NG Tensor to TF Tensor for this variable
   // Involves a copy from device to host
   // Returns the number of tensor copies made (0 or 1)
-  int copy_ng_to_tf() {
-    if (ng_tf_share_buffer_) {
-      return 0;
-    }
-    ReadNGTensor(ng_tensor_, &tf_tensor_);
-    return 1;
-  }
+  int copy_ng_to_tf();
 
   // Copies the TF Tensor to NG Tensor for this variable
   // Involves a copy from host to device
   // Returns the number of tensor copies made (0 or 1)
-  int copy_tf_to_ng() {
-    if (ng_tf_share_buffer_) {
-      return 0;
-    }
-    WriteNGTensor(ng_tensor_, &tf_tensor_);
-    return 1;
-  }
+  int copy_tf_to_ng();
 
   // updates the NGTensor with the new value
   // This new_value could be from ngraph-tensor, for e.g. when computed from
   // NGraphEncapsulateOp
   // and saved in Catalog
   // Returns the number of tensor copies made (0 or 1)
-  int update_ng_tensor(shared_ptr<ngraph::runtime::Tensor> new_value) {
-    ng_tensor_->copy_from(*new_value);
-    return 0;
-  }
+  int update_ng_tensor(shared_ptr<ngraph::runtime::Tensor> new_value);
 
   // updates the NGTensor with the new value
   // This new_value could be from tf-tensor, for e.g. when computed from a TF op
   // Returns the number of tensor copies made (0 or 1)
-  int update_ng_tensor(Tensor* new_value) {
-    WriteNGTensor(ng_tensor_, new_value);
-    if (ng_tf_share_buffer_) {
-      return 0;
-    }
-
-    return 1;
-  }
+  int update_ng_tensor(Tensor* new_value);
 
  private:
   mutex mu_;
